Details about the case.
This is a two piece case for a V4 Paintbrush PCB used for the ARTSEY.IO style of typing.  Files and more information for the Paintbrush can be found here: https://github.com/artseyio/thepaintbrush

The case files above are designed to be used with the right hand orientation of the keyboard.  The 'top tall' file is meant to be printed if you are using an OLED screen, and the 'top' is slightly lower if you are only using an MCU without the OLED.  The bottom will be the same for both the 'top tall' and the 'top' half of the case.  

If you are using a left handed keyboard orientation, use the bottom stl and use the 'no logo top tall' stl, and mirror it in your slicer.  

This case uses 4, 8mm x 2mm 3M bumpons, placed in the recesses in the bottom, as well as M2 heat inserts with M2x4mm screws to secure the PCB into the bottom half of the case.

Bumpons found here (F0802):  https://www.aliexpress.com/item/4001297404341.html
M2x4x3.2mm heat inserts found here: https://www.aliexpress.com/item/1005002986258961.html